Backend/PythonServer/requirements.txt
django
djangorestframework
django-cors-headers
pymysql
pyjwt
django-environ